Create add_daily_event.sql
The daily_usage event is included in the full schema but having it separate eases workflow and insures history is not lost when/if the event is updated.

To update the daily_usage event edit this file and do `mariadb aslusage < add_daily_event.sql

Be sure to update the full schema with `mysqldump --no-data alsusage >  aslusage.sql` if any changes are made to the daily_usage event.